requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
pydantic==1.10.17
python-dotenv==0.19.0
asyncpg==0.29.0
paho-mqtt==1.5.1
sse-starlette==0.7.2
pytest==6.2.5
httpx==0.19.0
aiosqlite==0.20.0